AAD sync listen / process inbound queue records
 *  Listen and pull events from the enbound queue.
 *  Bugfix in rbac-provider-ldap: returning string instead of int for getenv.
 *  Use python 3.5 image in rbac-provider-azure instead of 3.7.
 *  Bugfix in rbac-provider-azure: remove 3.6 f-strings.
 *  Bugfix multiple: name change of inbound and outbound queues.

[Task: 747]


refactor inbound filter tests and create outbound tests


add test to outbound user filter


add outbound group filter test


add invalid provider tests for group and user filter function


add test for outbound user creation


add invalid user and invalid provider test for outbound user creation


add all outbound filter tests to 98% coverage


refactor inbound filter tests and create outbound tests


add test to outbound user filter


add outbound group filter test


add invalid provider tests for group and user filter function


add test for outbound user creation


add invalid user and invalid provider test for outbound user creation


add all outbound filter tests to 98% coverage


refactor inbound filter tests and create outbound tests


add test to outbound user filter


add outbound group filter test


add invalid provider tests for group and user filter function


add test for outbound user creation


add invalid user and invalid provider test for outbound user creation


add all outbound filter tests to 98% coverage


refactor inbound filter tests and create outbound tests


add test to outbound user filter


add outbound group filter test


add invalid provider tests for group and user filter function


add test for outbound user creation


add invalid user and invalid provider test for outbound user creation


add all outbound filter tests to 98% coverage

+++ rbac/providers/ldap/delta_inbound_sync.py
@@ -34,7 +34,7 @@ LDAP_DC = os.getenv("LDAP_DC")
 LDAP_SERVER = os.getenv("LDAP_SERVER")
 LDAP_USER = os.getenv("LDAP_USER")
 LDAP_PASS = os.getenv("LDAP_PASS")
-DELTA_SYNC_INTERVAL_SECONDS = float(os.getenv("DELTA_SYNC_INTERVAL_SECONDS", "3600"))
+DELTA_SYNC_INTERVAL_SECONDS = int(os.getenv("DELTA_SYNC_INTERVAL_SECONDS", "3600"))
 
 LDAP_FILTER_USER_DELTA = "(&(objectClass=person)(whenChanged>=%s))"
 LDAP_FILTER_GROUP_DELTA = "(&(objectClass=group)(whenChanged>=%s))"
